backend/requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase & Auth
supabase==2.3.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
PyJWT==2.8.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# HTTP Client
httpx==0.26.0

# Astro Calculations
pyswisseph==2.10.3.2

# Utilities
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
